aider: Add is-active class to navbar items based on the current page.
# Aider chat conversation:

USER: with the navbar I need it to use is-active if the user is currently looking at this page, how do OI do this?

ASSISTANT: To add the `is-active` class to the navbar items based on the current page, we need to use Zola's built-in `page` variable to check the current URL and conditionally add the `is-active` class.

Here is the *SEARCH/REPLACE* block for `templates/base.html`:

templates/base.html
```python
<<<<<<< SEARCH
              <a class="navbar-item" href="/"> Home </a>

              <a class="navbar-item" href="/about"> About </a>

              <a class="navbar-item" href="/contact"> Contact </a>
=======
              <a class="navbar-item {% if page.url == '/' %}is-active{% endif %}" href="/"> Home </a>

              <a class="navbar-item {% if page.url == '/about/' %}is-active{% endif %}" href="/about"> About </a>

              <a class="navbar-item {% if page.url == '/contact/' %}is-active{% endif %}" href="/contact"> Contact </a>
>>>>>>> REPLACE
```