aider: Remove the `content` class from the div wrapping the blog index content.
# Aider chat conversation:

USER: hmm, I think because it's in a div with class content it's ignoring the bulma classes, recommendations?

ASSISTANT: To ensure that the Bulma classes are applied correctly, we can remove the `content` class from the div wrapping the blog index content. This will prevent any potential conflicts with Bulma's styling.

Here is the *SEARCH/REPLACE* block for `templates/blog_index.html`:

templates/blog_index.html
```html
<<<<<<< SEARCH
<div class="container">
    <h1 class="title">{{ page.title }}</h1>

    <div class="content">
        <ul>
        {% for section in sections %}
            {% for page in section.pages %}
            <li>
                <a class="has-text-link" href="{{ page.permalink }}">{{ page.title }}</a><br>
                <small class="has-text-grey">{{ page.date | date(format="long") }}</small>
            </li>
            {% endfor %}
        {% endfor %}
        </ul>
    </div>
</div>
=======
<div class="container">
    <h1 class="title">{{ page.title }}</h1>

    <ul>
    {% for section in sections %}
        {% for page in section.pages %}
        <li>
            <a class="has-text-link" href="{{ page.permalink }}">{{ page.title }}</a><br>
            <small class="has-text-grey">{{ page.date | date(format="long") }}</small>
        </li>
        {% endfor %}
    {% endfor %}
    </ul>
</div>
>>>>>>> REPLACE
```

This change will ensure that the Bulma classes are applied correctly to the blog index content.